[refactor] Move focus metrics from acq/align/autofocus.py to avoid circular import
MeasureSpotsFocus will need to be re-used in the fastem-calibrations code, but because that gets imported in odemis/acq/align/fastem.py it is causing a circular import.
Since the functions are more utility functions of the autofocus code, I moved them to odemis/util/focus.py
